backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]
groq
python-dotenv
psycopg2-binary
pydantic
pydantic-settings
requests
httpx
pytest
